Task 7 - Multi-Agent Newsletter Generator/requirements.txt
google-generativeai
pydantic
python-telegram-bot==20.0
beautifulsoup4
requests
numpy
scikit-learn
schedule
tqdm
python-dotenv
feedparser
langchain-community